Reset Naxxramas scripts to stock
