--- src/server/scripts/Northrend/Naxxramas/boss_anubrekhan.h
@@ -1,165 +1,0 @@
-#ifndef BOSS_ANUBREKHAN_H_
-#define BOSS_ANUBREKHAN_H_
-
-#include "ScriptMgr.h"
-#include "ScriptedCreature.h"
-#include "SpellInfo.h"
-#include "naxxramas.h"
-
-namespace Anubrekhan {
-
-enum Says
-{
-    SAY_AGGRO                       = 0,
-    SAY_GREET                       = 1,
-    SAY_SLAY                        = 2,
-    EMOTE_LOCUST                    = 3
-};
-
-enum GuardSays
-{
-    EMOTE_SPAWN                     = 1,
-    EMOTE_SCARAB                    = 2
-};
-
-enum Spells
-{
-    SPELL_IMPALE                    = 28783,
-    SPELL_LOCUST_SWARM              = 28785,
-    SPELL_SUMMON_CORPSE_SCRABS_5    = 29105,
-    SPELL_SUMMON_CORPSE_SCRABS_10   = 28864,
-    SPELL_BERSERK                   = 26662
-};
-
-enum Misc
-{
-    NPC_CORPSE_SCARAB               = 16698,
-    NPC_CRYPT_GUARD                 = 16573,
-
-    ACHIEV_TIMED_START_EVENT        = 9891
-};
-
-class boss_anubrekhan : public CreatureScript
-{
-public:
-    boss_anubrekhan() : CreatureScript("boss_anubrekhan") { }
-
-    CreatureAI* GetAI(Creature* pCreature) const override
-    {
-        return GetNaxxramasAI<boss_anubrekhanAI>(pCreature);
-    }
-
-    struct boss_anubrekhanAI : public BossAI
-    {
-        boss_anubrekhanAI(Creature* c) : BossAI(c, BOSS_ANUB)
-        {
-            sayGreet = false;
-        }
-
-        void SummonCryptGuards()
-        {
-            if (Is25ManRaid())
-            {
-                me->SummonCreature(NPC_CRYPT_GUARD, 3299.732f, -3502.489f, 287.077f, 2.378f, TEMPSUMMON_CORPSE_TIMED_DESPAWN, 60000);
-                me->SummonCreature(NPC_CRYPT_GUARD, 3299.086f, -3450.929f, 287.077f, 3.999f, TEMPSUMMON_CORPSE_TIMED_DESPAWN, 60000);
-            }
-        }
-
-        void Reset() override
-        {
-            BossAI::Reset();
-            SummonCryptGuards();
-            me->m_Events.KillAllEvents(false);
-        }
-
-        void JustSummoned(Creature* cr) override
-        {
-            if (me->IsInCombat())
-            {
-                cr->SetInCombatWithZone();
-                if (cr->GetEntry() == NPC_CRYPT_GUARD)
-                    cr->AI()->Talk(EMOTE_SPAWN, me);
-            }
-            summons.Summon(cr);
-        }
-
-        void SummonedCreatureDies(Creature* cr, Unit*) override
-        {
-            if (cr->GetEntry() == NPC_CRYPT_GUARD)
-            {
-                cr->CastSpell(cr, SPELL_SUMMON_CORPSE_SCRABS_10, true, nullptr, nullptr, me->GetGUID());
-                cr->AI()->Talk(EMOTE_SCARAB);
-            }
-        }
-
-        void JustDied(Unit*  killer) override
-        {
-            BossAI::JustDied(killer);
-            instance->DoStartTimedAchievement(ACHIEVEMENT_TIMED_TYPE_EVENT, ACHIEV_TIMED_START_EVENT);
-        }
-
-        void KilledUnit(Unit* victim) override
-        {
-            if (!victim->IsPlayer())
-                return;
-
-            Talk(SAY_SLAY);
-            victim->CastSpell(victim, SPELL_SUMMON_CORPSE_SCRABS_5, true, nullptr, nullptr, me->GetGUID());
-            instance->StorePersistentData(PERSISTENT_DATA_IMMORTAL_FAIL, 1);
-        }
-
-        void JustEngagedWith(Unit* who) override
-        {
-            BossAI::JustEngagedWith(who);
-            me->CallForHelp(30.0f);
-            Talk(SAY_AGGRO);
-
-            if (!summons.HasEntry(NPC_CRYPT_GUARD))
-                SummonCryptGuards();
-            if (!Is25ManRaid())
-            {
-                me->m_Events.AddEventAtOffset([&]
-                {
-                    me->SummonCreature(NPC_CRYPT_GUARD, 3331.217f, -3476.607f, 287.074f, 3.269f, TEMPSUMMON_CORPSE_TIMED_DESPAWN, 60000);
-                }, Milliseconds(urand(15000, 20000)));
-            }
-
-            ScheduleTimedEvent(15s, [&] {
-                DoCastRandomTarget(SPELL_IMPALE);
-            }, 20s);
-
-            ScheduleTimedEvent(70s, 2min, [&] {
-                Talk(EMOTE_LOCUST);
-                DoCastSelf(SPELL_LOCUST_SWARM);
-
-                me->m_Events.AddEventAtOffset([&]
-                {
-                    me->SummonCreature(NPC_CRYPT_GUARD, 3331.217f, -3476.607f, 287.074f, 3.269f, TEMPSUMMON_CORPSE_TIMED_DESPAWN, 60000);
-                }, 3s);
-
-            }, 90s);
-
-            me->m_Events.AddEventAtOffset([&]
-            {
-                DoCastSelf(SPELL_BERSERK, true);
-            }, 10min);
-        }
-
-        void MoveInLineOfSight(Unit* who) override
-        {
-            if (!sayGreet && who->IsPlayer())
-            {
-                Talk(SAY_GREET);
-                sayGreet = true;
-            }
-            ScriptedAI::MoveInLineOfSight(who);
-        }
-
-    private:
-        bool sayGreet;
-    };
-};
-
-}
-
-#endif
