+++ src/server/scripts/Northrend/Naxxramas/boss_loatheb.cpp
@@ -15,16 +15,158 @@
  * with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "boss_loatheb.h"
 #include "CreatureScript.h"
 #include "ScriptedCreature.h"
 #include "naxxramas.h"
 
+enum Spells
+{
+    SPELL_NECROTIC_AURA                         = 55593,
+    SPELL_SUMMON_SPORE                          = 29234,
+    SPELL_DEATHBLOOM                            = 29865,
+    SPELL_INEVITABLE_DOOM                       = 29204,
+    SPELL_BERSERK                               = 26662
+};
 
-using namespace Loatheb;
+enum Events
+{
+    EVENT_NECROTIC_AURA                         = 1,
+    EVENT_DEATHBLOOM                            = 2,
+    EVENT_INEVITABLE_DOOM                       = 3,
+    EVENT_BERSERK                               = 4,
+    EVENT_SUMMON_SPORE                          = 5,
+    EVENT_NECROTIC_AURA_FADING                  = 6,
+    EVENT_NECROTIC_AURA_REMOVED                 = 7
+};
 
-// no custom changes has been made for mod-playerbot other then placing
-// the impl in a header file
+enum Texts
+{
+    SAY_NECROTIC_AURA_APPLIED                   = 0,
+    SAY_NECROTIC_AURA_REMOVED                   = 1,
+    SAY_NECROTIC_AURA_FADING                    = 2
+};
+
+class boss_loatheb : public CreatureScript
+{
+public:
+    boss_loatheb() : CreatureScript("boss_loatheb") { }
+
+    CreatureAI* GetAI(Creature* pCreature) const override
+    {
+        return GetNaxxramasAI<boss_loathebAI>(pCreature);
+    }
+
+    struct boss_loathebAI : public BossAI
+    {
+        explicit boss_loathebAI(Creature* c) : BossAI(c, BOSS_LOATHEB), summons(me)
+        {
+            me->SetHomePosition(me->GetPositionX(), me->GetPositionY(), me->GetPositionZ(), me->GetOrientation());
+        }
+
+        uint8 doomCounter;
+        EventMap events;
+        SummonList summons;
+
+        void Reset() override
+        {
+            BossAI::Reset();
+            events.Reset();
+            summons.DespawnAll();
+            doomCounter = 0;
+        }
+
+        void JustSummoned(Creature* cr) override
+        {
+            cr->SetInCombatWithZone();
+            summons.Summon(cr);
+        }
+
+        void SummonedCreatureDies(Creature*  /*cr*/, Unit*) override
+        {
+            instance->SetData(DATA_SPORE_KILLED, 0);
+        }
+
+        void KilledUnit(Unit* who) override
+        {
+            if (who->IsPlayer())
+                instance->StorePersistentData(PERSISTENT_DATA_IMMORTAL_FAIL, 1);
+        }
+
+        void JustEngagedWith(Unit* who) override
+        {
+            BossAI::JustEngagedWith(who);
+            me->SetInCombatWithZone();
+            events.ScheduleEvent(EVENT_NECROTIC_AURA, 10s);
+            events.ScheduleEvent(EVENT_DEATHBLOOM, 5s);
+            events.ScheduleEvent(EVENT_INEVITABLE_DOOM, 2min);
+            events.ScheduleEvent(EVENT_SUMMON_SPORE, 15s);
+            events.ScheduleEvent(EVENT_BERSERK, 12min);
+        }
+
+        void JustDied(Unit* killer) override
+        {
+            BossAI::JustDied(killer);
+            summons.DespawnAll();
+        }
+
+        void UpdateAI(uint32 diff) override
+        {
+            if (!UpdateVictim() || !IsInRoom())
+                return;
+
+            events.Update(diff);
+            if (me->HasUnitState(UNIT_STATE_CASTING))
+                return;
+
+            switch (events.ExecuteEvent())
+            {
+                case EVENT_SUMMON_SPORE:
+                    me->CastSpell(me, SPELL_SUMMON_SPORE, true);
+                    events.Repeat(35s);
+                    break;
+                case EVENT_NECROTIC_AURA:
+                    me->CastSpell(me, SPELL_NECROTIC_AURA, true);
+                    Talk(SAY_NECROTIC_AURA_APPLIED);
+                    events.ScheduleEvent(EVENT_NECROTIC_AURA_FADING, 14s);
+                    events.ScheduleEvent(EVENT_NECROTIC_AURA_REMOVED, 17s);
+                    events.Repeat(20s);
+                    break;
+                case EVENT_DEATHBLOOM:
+                    me->CastSpell(me, SPELL_DEATHBLOOM, false);
+                    events.Repeat(30s);
+                    break;
+                case EVENT_INEVITABLE_DOOM:
+                    me->CastSpell(me, SPELL_INEVITABLE_DOOM, false);
+                    doomCounter++;
+                    events.Repeat(doomCounter < 6 ? 30s : 15s);
+                    break;
+                case EVENT_BERSERK:
+                    me->CastSpell(me, SPELL_BERSERK, true);
+                    break;
+                case EVENT_NECROTIC_AURA_FADING:
+                    Talk(SAY_NECROTIC_AURA_FADING);
+                    break;
+                case EVENT_NECROTIC_AURA_REMOVED:
+                    Talk(SAY_NECROTIC_AURA_REMOVED);
+                    break;
+            }
+            DoMeleeAttackIfReady();
+        }
+
+        bool IsInRoom()
+        {
+            // Calculate the distance between his home position to the gate
+            if (me->GetExactDist(me->GetHomePosition().GetPositionX(),
+                                 me->GetHomePosition().GetPositionY(),
+                                 me->GetHomePosition().GetPositionZ()) > 50.0f)
+            {
+                EnterEvadeMode();
+                return false;
+            }
+            return true;
+        }
+    };
+};
 
 void AddSC_boss_loatheb()
 {
